+++ src/client/view/components/Editor.js
