+++ src/client/view/components/Editor.scss
