+++ src/client/view/components/Icons.js
