+++ src/client/view/components/Prompt.scss
