+++ src/client/view/components/Search.scss
