+++ src/client/view/components/Sidebar.scss
